Shuffle tab files around + fix tidbit local bug
