+++ backend/active_statistics/server.py
@@ -11,8 +11,8 @@ from active_statistics.communication_schema import (
     RefreshStatusMessage,
 )
 from active_statistics.gui.gui import get_all_tabs, tab_tree
-from active_statistics.gui.tab_group import TabGroup
-from active_statistics.gui.tabs import Tab
+from active_statistics.tabs.tab_group import TabGroup
+from active_statistics.tabs.tabs import Tab
 from active_statistics.utils import human_readable_time, redis, rq
 from active_statistics.utils.environment_variables import evm
 from active_statistics.utils.local_storage import (
@@ -23,7 +23,7 @@ from active_statistics.utils.local_storage import (
 )
 from active_statistics.utils.routes import unauthorized_if_no_session_cookie
 from active_statistics.utils.sentry import set_up_sentry_for_server
-from flask import Flask, jsonify, make_response, redirect, request, session, url_for
+from flask import Flask, jsonify, make_response, redirect, request, session
 from requests.exceptions import HTTPError
 from rq.job import JobStatus
 from stravalib.client import Client
