+++ backend/active_statistics/tabs/image_tab.py
@@ -3,7 +3,7 @@ import os
 import shutil
 from typing import Any, Callable, Iterator
 
-from active_statistics.gui.tabs import Tab
+from active_statistics.tabs.tabs import Tab
 from active_statistics.utils.environment_variables import EnvironmentVariableManager
 from active_statistics.utils.local_storage import (
     get_activity_iterator,
